removed redundant code - improved error checking for player
