--- app/(root)/dashboard/videos/[name]/watch/not-found.tsx
@@ -1,22 +1,0 @@
-import Link from "next/link";
-import { Frown } from "lucide-react";
-
-export default function NotFound() {
-  return (
-    <main className="flex h-full flex-col items-center justify-center gap-2">
-      <Frown className="w-10 text-gray-400" />
-      <h2 className="text-xl font-semibold">404 Not Found</h2>
-      <p>Could not find the requested video.</p>
-      <p>There is a good chance the page has refreshed and the URL lost.</p>
-      <p>
-        You will need to go back and rescan the folder where your video are.
-      </p>
-      <Link
-        href="/dashboard/"
-        className="mt-4 rounded-md bg-blue-500 px-4 py-2 text-sm text-white transition-colors hover:bg-blue-400"
-      >
-        Go Back
-      </Link>
-    </main>
-  );
-}
