--- app/(root)/dashboard/videos/page.tsx
@@ -1,3 +1,0 @@
-export default function Videos() {
-  return <h1>Videos</h1>;
-}
