+++ app/(root)/dashboard/watch/[name]/layout.tsx
